Add File Uploader
Add a simple file uploader component which prompts the user to choose a file.
Once a file is chosen, the user can upload it which will then trigger a mock
api call while polling for file upload progress. The file and status will then
be presented as a list item within the screen. The user can also cancel
the upload which will then remove it from the uploadList.